+++ src/main/java/com/capstone/autism_training/deck/DeckModel.java
@@ -2,12 +2,12 @@ package com.capstone.autism_training.deck;
 
 public class DeckModel {
 
-    public int id;
+    public long id;
     public byte[] image;
     public String name;
     public String description;
 
-    public DeckModel(int id, byte[] image, String name, String description) {
+    public DeckModel(long id, byte[] image, String name, String description) {
         this.id = id;
         this.image = image;
         this.name = name;
